requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.3
Frozen-Flask==1.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
PyYAML==6.0.3
Werkzeug==3.1.6
